+++ src/ImageFitter/ImageFitter.h
@@ -26,6 +26,8 @@ struct FitData {
     size_t n_notnan; // number of pixels excluding nan pixels
     size_t offset_x;
     size_t offset_y;
+    std::vector<int> fit_values_indexes;
+    std::vector<double> initial_values;
 };
 
 struct FitStatus {
@@ -39,7 +41,7 @@ class ImageFitter {
 public:
     ImageFitter();
     bool FitImage(size_t width, size_t height, float* image, const std::vector<CARTA::GaussianComponent>& initial_values,
-        CARTA::FittingResponse& fitting_response, size_t offset_x = 0, size_t offset_y = 0);
+        const std::vector<bool>& fixed_params, CARTA::FittingResponse& fitting_response, size_t offset_x = 0, size_t offset_y = 0);
 
 private:
     FitData _fit_data;
@@ -51,7 +53,7 @@ private:
     const size_t _max_iter = 200;
 
     void CalculateNanNum();
-    void SetInitialValues(const std::vector<CARTA::GaussianComponent>& initial_values);
+    void SetInitialValues(const std::vector<CARTA::GaussianComponent>& initial_values, const std::vector<bool>& fixed_params);
     int SolveSystem();
     void SetResults();
     std::string GetLog();
@@ -59,7 +61,9 @@ private:
     static int FuncF(const gsl_vector* fit_params, void* fit_data, gsl_vector* f);
     static void Callback(const size_t iter, void* params, const gsl_multifit_nlinear_workspace* w);
     static void ErrorHandler(const char* reason, const char* file, int line, int gsl_errno);
-    static CARTA::GaussianComponent GetGaussianComponent(gsl_vector* value_vector, size_t index);
+    static std::tuple<double, double, double, double, double, double> GetGaussianParams(const gsl_vector* value_vector, size_t index,
+        std::vector<int>& fit_values_indexes, std::vector<double>& initial_values, size_t offset_x = 0, size_t offset_y = 0);
+    static CARTA::GaussianComponent GetGaussianComponent(std::tuple<double, double, double, double, double, double> params);
 };
 
 } // namespace carta